requirements.txt
python-dotenv
streamlit
tiktoken
os
langchain
openai
pinecone